# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * El Hawes Travel Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface Trip {
  id: number;
  title: string;
  slug: string;
  description: string;
  shortDescription: string;
  location: string;
  country: string;
  durationDays: number;
  priceDzd: number;
  /** @nullable */
  originalPriceDzd: number | null;
  maxGroupSize: number;
  spotsLeft: number;
  imageUrl: string;
  galleryImages: string[];
  highlights: string[];
  included: string[];
  excluded: string[];
  isFeatured: boolean;
  isActive: boolean;
  /** @nullable */
  departureDate: string | null;
  /** @nullable */
  returnDate: string | null;
  difficulty: string;
  category: string;
  rating: number;
  reviewCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTripBody {
  title: string;
  slug: string;
  description: string;
  shortDescription: string;
  location: string;
  country: string;
  durationDays: number;
  priceDzd: number;
  /** @nullable */
  originalPriceDzd?: number | null;
  maxGroupSize: number;
  spotsLeft: number;
  imageUrl: string;
  galleryImages?: string[];
  highlights?: string[];
  included?: string[];
  excluded?: string[];
  isFeatured?: boolean;
  isActive?: boolean;
  /** @nullable */
  departureDate?: string | null;
  /** @nullable */
  returnDate?: string | null;
  difficulty?: string;
  category?: string;
}

export interface UpdateTripBody {
  title?: string;
  slug?: string;
  description?: string;
  shortDescription?: string;
  location?: string;
  country?: string;
  durationDays?: number;
  priceDzd?: number;
  /** @nullable */
  originalPriceDzd?: number | null;
  maxGroupSize?: number;
  spotsLeft?: number;
  imageUrl?: string;
  galleryImages?: string[];
  highlights?: string[];
  included?: string[];
  excluded?: string[];
  isFeatured?: boolean;
  isActive?: boolean;
  /** @nullable */
  departureDate?: string | null;
  /** @nullable */
  returnDate?: string | null;
  difficulty?: string;
  category?: string;
}

export interface Booking {
  id: number;
  tripId: number;
  tripTitle: string;
  fullName: string;
  email: string;
  phone: string;
  numberOfPeople: number;
  /** @nullable */
  message: string | null;
  status: string;
  totalPriceDzd: number;
  /** @nullable */
  preferredDate: string | null;
  isContacted: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBookingBody {
  tripId: number;
  fullName: string;
  email: string;
  phone: string;
  numberOfPeople: number;
  message?: string;
  preferredDate?: string;
}

export interface UpdateBookingBody {
  status?: string;
  isContacted?: boolean;
}

export interface InstagramPost {
  id: string;
  imageUrl: string;
  /** @nullable */
  caption: string | null;
  /** @nullable */
  location: string | null;
  likes: number;
  permalink: string;
}

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface AdminAuthResponse {
  authenticated: boolean;
  username: string;
}

export interface DashboardStats {
  totalTrips: number;
  activeTrips: number;
  totalBookings: number;
  pendingBookings: number;
  confirmedBookings: number;
  totalRevenueDzd: number;
  recentBookings: Booking[];
}

export type ListTripsParams = {
  featured?: string;
  limit?: number;
};

export type GetInstagramFeedParams = {
  limit?: number;
};